+++ crates/types/src/field.rs
@@ -78,6 +78,7 @@ pub enum IdentityField {
 pub enum PrincipalField {
     Archive = ARCHIVE_FIELD,
     ParticipantIdentities = 45,
+    IdentityAddresses = 46,
     DefaultCalendarId = 47,
     DefaultAddressBookId = 48,
     ActiveScriptId = 49,
@@ -156,6 +157,7 @@ impl From<PrincipalField> for u8 {
     fn from(value: PrincipalField) -> Self {
         match value {
             PrincipalField::ParticipantIdentities => 45,
+            PrincipalField::IdentityAddresses => 46,
             PrincipalField::DefaultCalendarId => 47,
             PrincipalField::DefaultAddressBookId => 48,
             PrincipalField::ActiveScriptId => 49,